move test classes to correct pacakge.
